aider: Update openai_command to use the content from the last 'user' message in the incoming query.
# Aider chat conversation:

USER: make sure to take the content from the last entry in the incoming query to chatcompletion

ASSISTANT: To ensure that the content from the last entry in the incoming query is taken, we need to modify the `openai_command` function in the `app.py` file.

Currently, the function is using the `next` function to get the first message with the role 'user'. We need to change this to get the last message with the role 'user' instead.

Here is the *SEARCH/REPLACE* block:

app.py
```python
<<<<<<< SEARCH
command_str = next((msg['content'] for msg in messages if msg['role'] == 'user'), None)
=======
command_str = next((msg['content'] for msg in reversed(messages) if msg['role'] == 'user'), None)
>>>>>>> REPLACE
```

This change will ensure that the `openai_command` function uses the content from the last 'user' message in the incoming query.