aider: Add section on running the Docker container in `README.md`.
# Aider chat conversation:

USER: modify `README.md` to include a section on running the docker

ASSISTANT: To make this change we need to modify `README.md` to:

1. Add a section on running the Docker container.

Here is the *SEARCH/REPLACE* block:

README.md
<source>
<<<<<<< SEARCH
## Running the Application

To run this application, you need to have Docker installed on your machine. Follow the instructions below:

1. Build the Docker image:
2. Run the Docker container:
=======
## Running the Application

To run this application, you need to have Docker installed on your machine. Follow the instructions below:

1. Build the Docker image: `docker build -t tmux-api .`
2. Run the Docker container: `docker run -p 5000:5000 tmux-api`
>>>>>>> REPLACE
</source>